+++ ckks/params.go
@@ -307,12 +307,21 @@ var (
 // fields and is used to express unchecked user-defined parameters literally into
 // Go programs. The NewParametersFromLiteral function is used to generate the actual
 // checked parameters from the literal representation.
+//
+// Users must set the polynomial degree (in log_2, LogN) and the coefficient modulus, by either setting
+// the Q and P fields to the desired moduli chain, or by setting the LogQ and LogP fields to
+// the desired moduli sizes (in log_2). Users must also specify a default initial scale for the plaintexts.
+//
+// Optionally, users may specify the error variance (Sigma), the secrets' density (H), the ring
+// type (RingType) and the number of slots (in log_2, LogSlots). If left unset, standard default values for
+// these field are substituted at parameter creation (see NewParametersFromLiteral).
 type ParametersLiteral struct {
 	LogN         int // Ring degree (power of 2)
 	Q            []uint64
 	P            []uint64
-	LogQ         []int   `json:",omitempty"`
-	LogP         []int   `json:",omitempty"`
+	LogQ         []int `json:",omitempty"`
+	LogP         []int `json:",omitempty"`
+	H            int
 	Sigma        float64 // Gaussian sampling variance
 	LogSlots     int
 	DefaultScale float64
@@ -356,8 +365,13 @@ func NewParameters(rlweParams rlwe.Parameters, logSlots int, defaultScale float6
 
 // NewParametersFromLiteral instantiate a set of CKKS parameters from a ParametersLiteral specification.
 // It returns the empty parameters Parameters{} and a non-nil error if the specified parameters are invalid.
+//
+// If the `LogSlots` field is left unset, its value is set to `LogN-1` for the Standard ring and to `LogN` for
+// the conjugate-invariant ring.
+//
+// See `rlwe.NewParametersFromLiteral` for default values of the other optional fields.
 func NewParametersFromLiteral(pl ParametersLiteral) (Parameters, error) {
-	rlweParams, err := rlwe.NewParametersFromLiteral(rlwe.ParametersLiteral{LogN: pl.LogN, Q: pl.Q, P: pl.P, LogQ: pl.LogQ, LogP: pl.LogP, Sigma: pl.Sigma, RingType: pl.RingType})
+	rlweParams, err := rlwe.NewParametersFromLiteral(rlwe.ParametersLiteral{LogN: pl.LogN, Q: pl.Q, P: pl.P, LogQ: pl.LogQ, LogP: pl.LogP, H: pl.H, Sigma: pl.Sigma, RingType: pl.RingType})
 	if err != nil {
 		return Parameters{}, err
 	}
@@ -532,6 +546,10 @@ func (p Parameters) Equals(other Parameters) bool {
 }
 
 // CopyNew makes a deep copy of the receiver and returns it.
+//
+// Deprecated: Parameter is now a read-only struct, except for the UnmarshalBinary method: deep copying should only be
+// required to save a Parameter struct before calling its UnmarshalBinary method and it will be deprecated when
+// transitioning to a immutable serialization interface.
 func (p Parameters) CopyNew() Parameters {
 	p.Parameters = p.Parameters.CopyNew()
 	return p
@@ -577,7 +595,7 @@ func (p Parameters) MarshalBinarySize() int {
 
 // MarshalJSON returns a JSON representation of this parameter set. See `Marshal` from the `encoding/json` package.
 func (p Parameters) MarshalJSON() ([]byte, error) {
-	return json.Marshal(ParametersLiteral{LogN: p.LogN(), Q: p.Q(), P: p.P(), Sigma: p.Sigma(), LogSlots: p.logSlots, DefaultScale: p.defaultScale, RingType: p.RingType()})
+	return json.Marshal(ParametersLiteral{LogN: p.LogN(), Q: p.Q(), P: p.P(), H: p.HammingWeight(), Sigma: p.Sigma(), LogSlots: p.logSlots, DefaultScale: p.defaultScale, RingType: p.RingType()})
 }
 
 // UnmarshalJSON reads a JSON representation of a parameter set into the receiver Parameter. See `Unmarshal` from the `encoding/json` package.